+++ lib/oban/queue/producer.ex
@@ -3,7 +3,7 @@ defmodule Oban.Queue.Producer do
 
   use GenServer
 
-  import Oban.Breaker, only: [open_circuit: 1, trip_errors: 0, trip_circuit: 2]
+  import Oban.Breaker, only: [open_circuit: 1, trip_errors: 0, trip_circuit: 3]
   import Oban.Notifier, only: [insert: 0, signal: 0]
 
   alias Oban.{Config, Notifier, Query}
@@ -195,7 +195,7 @@ defmodule Oban.Queue.Producer do
 
     state
   rescue
-    exception in trip_errors() -> trip_circuit(exception, state)
+    exception in trip_errors() -> trip_circuit(exception, __STACKTRACE__, state)
   end
 
   defp start_listener(%State{conf: conf} = state) do
@@ -229,7 +229,7 @@ defmodule Oban.Queue.Producer do
 
     state
   rescue
-    exception in trip_errors() -> trip_circuit(exception, state)
+    exception in trip_errors() -> trip_circuit(exception, __STACKTRACE__, state)
   end
 
   defp pulse(%State{circuit: :disabled} = state) do
@@ -249,7 +249,7 @@ defmodule Oban.Queue.Producer do
 
     state
   rescue
-    exception in trip_errors() -> trip_circuit(exception, state)
+    exception in trip_errors() -> trip_circuit(exception, __STACKTRACE__, state)
   end
 
   defp dispatch(%State{paused: true} = state) do
@@ -289,7 +289,7 @@ defmodule Oban.Queue.Producer do
         {:noreply, state}
     end
   rescue
-    exception in trip_errors() -> {:noreply, trip_circuit(exception, state)}
+    exception in trip_errors() -> {:noreply, trip_circuit(exception, __STACKTRACE__, state)}
   end
 
   defp dispatch_now?(%State{dispatched_at: nil}), do: true